+++ lib/widgets/filter_grids/common/action_delegates/chip_set.dart
@@ -28,7 +28,7 @@ import 'package:aves/widgets/dialogs/aves_dialog.dart';
 import 'package:aves/widgets/dialogs/filter_editors/cover_selection_dialog.dart';
 import 'package:aves/widgets/dialogs/tile_view_dialog.dart';
 import 'package:aves/widgets/map/map_page.dart';
-import 'package:aves/widgets/search/search_delegate.dart';
+import 'package:aves/widgets/search/collection_search_delegate.dart';
 import 'package:aves/widgets/stats/stats_page.dart';
 import 'package:aves/widgets/viewer/slideshow_page.dart';
 import 'package:aves_model/aves_model.dart';
@@ -388,6 +388,8 @@ abstract class ChipSetActionDelegate<T extends CollectionFilter> with FeedbackMi
     if (confirmed == null || !confirmed) return;
 
     final filters = getSelectedFilters(context);
+    if (!await unlockFilters(context, filters)) return;
+
     settings.changeFilterVisibility(filters, false);
 
     browse(context);